Finally produced an image that works.
The key is to generate the C array with 1-bit color depth, 8 pixels per byte,
pixels arranged horizontally on the byte.  The image had to be mirrored
horizontally as well.

The tool that finally worked for me was LCDAssistant
http://en.radzio.dxp.pl/bitmap_converter/ which is only for Windows but works
fine on Linux with wine.

And imagemagick/gimp for the preprocessing and mirroring.